splits seed
